ShowcaseTypeAdapter refactoring.
