+++ src/main/java/com/yandex/money/api/typeadapters/AllowedMoneySourceTypeAdapter.java
@@ -1,0 +1,64 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2015 NBCO Yandex.Money LLC
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package com.yandex.money.api.typeadapters;
+
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonParseException;
+import com.google.gson.JsonPrimitive;
+import com.google.gson.JsonSerializationContext;
+import com.yandex.money.api.model.AllowedMoneySource;
+
+import java.lang.reflect.Type;
+
+/**
+ * @author Anton Ermak (ermak@yamoney.ru)
+ */
+public final class AllowedMoneySourceTypeAdapter extends BaseTypeAdapter<AllowedMoneySource> {
+
+    public static final AllowedMoneySourceTypeAdapter INSTANCE =
+            new AllowedMoneySourceTypeAdapter();
+
+    private AllowedMoneySourceTypeAdapter() {
+    }
+
+    @Override
+    public AllowedMoneySource deserialize(JsonElement json, Type typeOfT,
+                                          JsonDeserializationContext context)
+            throws JsonParseException {
+        return AllowedMoneySource.parse(json.getAsJsonPrimitive().getAsString());
+    }
+
+    @Override
+    public JsonElement serialize(AllowedMoneySource src, Type typeOfSrc,
+                                 JsonSerializationContext context) {
+        return new JsonPrimitive(src.code);
+    }
+
+    @Override
+    protected Class<AllowedMoneySource> getType() {
+        return AllowedMoneySource.class;
+    }
+}
